Make a pool of workers to handle Analyses
